Dtos a Data Annotations concluido

+++ Backend/src/ProEventos.API/Controllers/EventosController.cs
@@ -23,7 +23,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -39,7 +39,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento por id não encontrado.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -55,7 +55,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Eventos por tema não encontrados.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -71,7 +71,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -87,7 +87,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar atualizar evento.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -102,10 +102,13 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var eventos = await _eventoService.GetAllEventosAsync(true);
+                if (eventos == null) return NoContent();
+
                 if (await _eventoService.DeleteEvento(id)) return Ok("Eliminado");
                 else
                 {
-                    return BadRequest("O evento não foi eliminado");
+                    throw new Exception("Ocorreu um problema não específico ao tentar eliminar o evento");
                 }
             }
             catch (Exception ex)
